requirements.txt
# Chotu AI Assistant - Python Dependencies

# Core AI and OpenAI
openai==1.3.7
python-dotenv==1.0.0

# Web Framework (MCP Server)
flask==2.3.3
flask-cors>=4.0.0

# Voice Recognition and Audio
speechrecognition==3.10.0
pyaudio==0.2.11
pyttsx3>=2.90

# Web Automation
selenium>=4.15.0
webdriver-manager>=4.0.0
beautifulsoup4>=4.12.0
requests==2.31.0

# Natural Language Processing
nltk>=3.8.1
textblob>=0.17.1

# System Integration
psutil>=5.9.0
pyobjc>=9.2.0

# Data Processing
pandas>=2.1.0
numpy>=1.24.0

# Utilities
python-dateutil>=2.8.0
pytz>=2023.3
rich>=13.0.0
click>=8.1.0

# Development and Testing
setuptools>=68.0.0
pytest>=7.4.0

# macOS Specific
osascript>=2020.12.3

# Memory and Storage
jsonschema>=4.19.0

# Logging
loguru>=0.7.0

# Security
cryptography>=41.0.0
